+++ examples/98_Supremacy/src/Components.h
@@ -144,8 +144,6 @@ struct VisualComponent
     iTexturePtr _texture;
     bool _castShadow;
     bool _scaleAnimation;
-    bool _hasShield;
-    bool _isEnraged;
     iaTime _timerOffset;
 };