+++ examples/98_Supremacy/src/GameLayer.h
@@ -46,8 +46,6 @@ struct EnemyClass
     float64 _speed; 
     uint32 _xpDrop;
     uint32 _coinDrop;
-    bool _hasShield;
-    bool _isEnraged;
 };
 
 /*! GameLayer