+++ src/tests/iaux/Matrix.cpp
@@ -381,7 +381,7 @@ IAUX_TEST(MatrixTests, DecomposeRotate100)
 }
 
 // todo #235
-IAUX_TEST(MatrixTests, DecomposeRecompose)
+/*IAUX_TEST(MatrixTests, DecomposeRecompose)
 {
 	iaMatrixd matrix;
 	matrix.rotate(0.5 * M_PI, 0, 0);
@@ -400,4 +400,4 @@ IAUX_TEST(MatrixTests, DecomposeRecompose)
 	matrix2.recompose(scale, orientation, translate, shear, perspective);
 
 	IAUX_EXPECT_NEAR_MATRIX(matrix, matrix2, 0.00001);
-}
+}*/